Port over dns_check tests to pytest
