+++ dns_check/tox.ini
@@ -1,0 +1,26 @@
+[tox]
+minversion = 2.0
+basepython = py27
+envlist =
+    dns_check
+    flake8
+
+[testenv]
+platform = linux|darwin|win32
+
+[testenv:http_check]
+deps =
+    ../datadog_checks_base
+    -rrequirements-dev.txt
+commands =
+    pip install --require-hashes -r requirements.txt
+    pytest -v
+
+[testenv:flake8]
+skip_install = true
+deps = flake8
+commands = flake8 .
+
+[flake8]
+exclude = .eggs,.tox
+max-line-length = 120
